calibration wrapped

+++ main.py
@@ -20,8 +20,11 @@ import argparse
 
 
 class UAVCalibration:
+    """Calibration parameters for the UAV.
+    
+    Body frame is in the same orientation as  the IMU frame but with an offset.
+    """
     def __init__(self):
-        # IMU to Camera transformation (T_ic)
         self.T_ci = np.array([
             [ 0.011106298412152327,  0.9999324199187616,  0.0034359468839849595, 0.036802732375442404],
             [-0.999832733821092,  0.01115499451474039,  -0.014493808237225339,  -0.008332238900780303],
@@ -29,15 +32,27 @@ class UAVCalibration:
             [ 0.        ,  0.        ,  0.        ,  1.        ],
         ])
         self.R_ci = self.T_ci[:3, :3]
-        self.R_ic = self.R_ci.T  # Rotation from Camera to IMU
-        self.t_imu2cam = self.T_ci[:3, 3]  # Translation from IMU to Camera in Camera frame
-        self.t_body2imu = np.array([-3.89588892, 0, -27.96108098])  # body to IMU in IMU frame
-        self.t_imu2body = np.array([11.0, 0.0, 26.0])  # IMU to body in body frame
-        
+        self.R_ic = self.R_ci.T  # Rotation: camera to IMU
+        self.t_imu2cam = self.T_ci[:3, 3]  # Translation from IMU to camera in camera frame
+        self.t_body2imu = np.array([-0.0389588892, 0, -0.2796108098])  # body to IMU in IMU frame
+        self.t_imu2body = -self.t_body2imu  # IMU to body in body frame
+        # self.t_imu2body = np.array([11.0, 0.0, 26.0])  # IMU to body in body frame
         self.t_body2cam = (self.T_ci @ np.hstack((self.t_body2imu, [1])).reshape(-1, 1)).flatten()[:3]  # body to camera in camera frame
+        self.t_body2cam_imu = (self.R_ic @ self.t_body2cam.reshape(-1, 1)).flatten()  # body to camera in IMU frame
+        # self.t_cam2body = -(self.R_ic @ self.t_imu2cam.reshape(-1, 1)).flatten()  # camera to body in IMU/body frame
+        self.t_cam2body = -self.t_body2cam_imu
         
-        # self.t_cam2body = 
-
+    def get_alt_cam(self, alt_body, R_wi):
+        """
+        Get the altitude of the camera in the world frame.
+        :param alt_body: Altitude in body frame.
+        :param R_wi: Rotation from world to IMU frame.
+        :return: Altitude of the camera in the world frame.
+        """
+        alt_offset = (R_wi @ self.t_body2cam_imu.reshape(-1, 1)).flatten()[-1]
+        assert alt_offset <= 0, "Drone upside down?"
+        alt_cam = alt_body - (R_wi @ self.t_body2cam_imu.reshape(-1, 1)).flatten()[-1]
+        return alt_cam
 
 
 if __name__ == '__main__':
@@ -99,13 +114,19 @@ if __name__ == '__main__':
         logger.info(f"Removed previous frames: {save_frames_to}")
     os.mkdir(save_frames_to)
 
-    if os.path.isfile(vid_pth):
-        os.remove(vid_pth)
+    if args.save_vid and os.path.isfile(vid_pth):
+            os.remove(vid_pth)
 
 
 
     calib = UAVCalibration()
     i = 0
+    confidence_threshold = 0.67
+    logger.info(f"YOLO confidence threshold: {confidence_threshold}")
+    
+    coords = []
+    bbox_log_path = os.path.join("logs/bbox_log.txt")
+    bbox_log_file = open(bbox_log_path, "w")
     for ts, frame, translation, R_wi, zone in image_stream(
         bag_pth, 
         frame_topic, 
@@ -115,7 +136,9 @@ if __name__ == '__main__':
         gps_topic=gps_topic, 
         imu_topic=imu_topic
     ):
-        result = model.predict(frame, verbose=False, conf=0.61)[0]
+        
+        
+        result = model.predict(frame, verbose=False, conf=confidence_threshold)[0]
         
         if args.loc:
             R_wc = R_wi @ calib.R_ic
@@ -136,24 +159,41 @@ if __name__ == '__main__':
                 # zone, easting, northing = LLtoUTM(23, lat_uav, long_uav)
                 northing, easting, alt_uav = translation  # NED
                 ray_cam = (cam.reproject(img_coord)).reshape(-1, 1)
-                ray_body =(R_cd.T @ (ray_cam)).flatten() # ray in body frame
+                ray_body =(calib.R_ic @ (ray_cam)).flatten() # ray in body frame
                 ray_world = (R_wc @ (ray_cam)).flatten()  # ray in world frame
                 
                 
-                cam_loc = translation + (R_wd @ t_cam2body.reshape(-1, 1)).flatten()  # camera: northing, easting, altitude
-                alt_cam = cam_loc[-1]
+                alt_cam = calib.get_alt_cam(alt_uav, R_wi)
                 s = -(alt_cam)/ray_world[-1]
                 
-                target2body = s * ray_body + t_cam2body
-                target2body_world = (R_wd @ target2body.reshape(-1, 1)).flatten()
+                
+                
+                target2body = s * ray_body + calib.t_cam2body
+                target2body_world = (R_wi @ target2body.reshape(-1, 1)).flatten()
                 
                 world_coord = translation + target2body_world
+                
+                # print(f"Alt_uav: {alt_uav}, Alt_cam: {alt_cam}")
+                # print(f"s: {s}")
+                # print(f"Ray Cam: {ray_cam.flatten()}")
+                # print(f"Ray World: {ray_world.flatten()}")
+                # print(f"Trans: {[northing, easting, alt_uav]}")
+                # print(f"World Coord: {[world_coord[0], world_coord[1], world_coord[2]]}")
+                # print("======================================================")
+                
+                
+                coords.append(world_coord)
+                
+                
+                # label_coord = world_coord - starting_coord
 
                 label = f"({np.round(world_coord - starting_coord, 2)})"
 
             else:
                 conf = box.conf.item()
                 label = f"Person {conf:.2f}"
+                # Log bounding box: frame_id, x1, y1, x2, y2, confidence
+                bbox_log_file.write(f"{i:06d},{xyxy[0]},{xyxy[1]},{xyxy[2]},{xyxy[3]},{conf:.4f}\n")
 
             # Draw bounding box
             cv2.rectangle(frame, tuple(xyxy[:2]), tuple(xyxy[2:]), box_color, thickness)
@@ -169,9 +209,12 @@ if __name__ == '__main__':
             cv2.putText(frame, label, (xyxy[0], xyxy[1] - 5),
                         cv2.FONT_HERSHEY_SIMPLEX, 0.8, (255, 255, 255), 2)
         i += 1
-        # Always save the frame, even if no person is found
-        cv2.imwrite(os.path.join(save_frames_to, f'frame_{i:06d}.jpg'), frame)
 
-    
+        if args.save_vid:
+            # Always save the frame, even if no person is found
+            cv2.imwrite(os.path.join(save_frames_to, f'frame_{i:06d}.jpg'), frame)
+        
+    bbox_log_file.close()
+    np.save('logs/dry_run_1_half.npy', coords)
     if args.save_vid:
         save_video(vid_pth, save_frames_to)
